Update: Modularise Totals component
